+++ csw-time-client/src/test/scala/csw/time/client/scheduling_spikes/TimerExample.java
@@ -37,8 +37,8 @@ class MyTask extends TimerTask {
         TimeSpec timeSpec = new TimeSpec();
 
         TimeLibrary.clock_gettime(0, timeSpec);
-        long s = timeSpec.seconds;
-        String n = String.format("%09d",timeSpec.nanoseconds);
+        long s = timeSpec.seconds.longValue();
+        String n = String.format("%09d",timeSpec.nanoseconds.longValue());
         buf.add(s+""+n);
